client/requirements.txt
anyio==3.6.1
asyncpg==0.26.0
click==8.1.3
fastapi==0.85.0
greenlet==1.1.3
h11==0.14.0
idna==3.4
psycopg2==2.9.3
pydantic==1.10.2
python-dotenv==0.21.0
sniffio==1.3.0
SQLAlchemy==1.4.41
starlette==0.20.4
typing_extensions==4.3.0
uvicorn==0.18.3
